Add `x11vnc` server integration tests
This is a small starter test suite which verifies basic connection and
auth with a local `x11vnc` server run in a sub-process. Obviously the
`x11vnc` program need to be installed locally for this suite to run.

At the time of writing this test suite exemplifies the following issues:
- https://github.com/barneygale/asyncvnc/issues/1
- https://github.com/barneygale/asyncvnc/issues/2

Further work to put this into CI is not included here but is highly
recommended ;)